// src/models/Auth/AuthVm.ts
/* eslint-disable prettier/prettier */
export class AuthVm {
  userid: string ='0000000000000';
  status: boolean = false;
  statuscode: number = 401;
  username: string ='noname';
  message: string= 'Không tồn tại';
  accessToken: string ='000';
  refreshToken: string ='000';
  role: string ='000';
}